Falta resolver el error infernal de la recursion

+++ crearApi/src/main/java/servlets/Asignaturas.java
@@ -6,9 +6,7 @@
 package servlets;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -17,10 +15,8 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import model.Alumno;
 import model.Asignatura;
 import model.Completado;
-import servicios.AlumnosServicios;
 import servicios.AsignaturasServicios;
 
 /**
@@ -68,42 +64,54 @@ public class Asignaturas extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         AsignaturasServicios as = new AsignaturasServicios();
-        Completado completado = new Completado();
         Asignatura asignatura = (Asignatura) request.getAttribute("asignatura");
         if (as.updateAsignatura(asignatura)) {
-            completado.setMensaje("Actualizado");
+            request.setAttribute("json", "Actualizado");
         } else {
-            completado.setMensaje("Se produjo un error al actualizar");
+            request.setAttribute("json", "Se produjo un error al actualizar");
         }
-        request.setAttribute("json", completado);
+        
     }
 
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         AsignaturasServicios as = new AsignaturasServicios();
-        Completado completado = new Completado();
+        
         Asignatura asignatura = (Asignatura) req.getAttribute("asignatura");
-        if (as.deleteAsignatura(asignatura)) {
-            completado.setMensaje("Borrado");
-            
-        } else {
-            resp.setStatus(500);
-            completado.setMensaje("Se produjo un error al borrar");
+        
+        if (!("ok").equals(req.getParameter("deletesiosi"))){
+            if (as.deleteAsignatura(asignatura)) {
+                resp.setStatus(200);
+                req.setAttribute("json", "Borrado correctamente");
+            } else {
+                resp.setStatus(500);
+                req.setAttribute("json", "Se produjo un error al borrar");
+            }
+        }else{
+            try {
+                if (as.completeDeleteAlumno(asignatura) == 1) {
+                    resp.setStatus(200);
+                    req.setAttribute("json", "Borrado correctamente");
+                } else {
+                    resp.setStatus(500);
+                    req.setAttribute("json", "Se produjo un error al borrar");
+                }
+            } catch (SQLException ex) {
+                Logger.getLogger(Alumnos.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
-        req.setAttribute("json", completado); //To change body of generated methods, choose Tools | Templates.
+         //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
     protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         AsignaturasServicios as = new AsignaturasServicios();
-        Completado completado = new Completado();
         Asignatura asignatura = (Asignatura) req.getAttribute("asignatura");
         if (as.addAsignatura(asignatura)) {
-            completado.setMensaje("Insertado");
+            req.setAttribute("json", "Insertado");
         } else {
-            completado.setMensaje("Se produjo un error al insertar");
+            req.setAttribute("json", "Se produjo un error al insertar");
         }
-        req.setAttribute("json", completado);
     }
 
     /**
